Create CODE_OF_CONDUCT.md
Candidates are expected to conduct themselves with the highest levels of honesty and integrity during examination and reminded that TCS have provision in place to deter dishonesty: 1.1 Exams are conducted using TCS owned and managed platforms. 1.2 Candidates will be observed during their assessments by Assessors and/or Invigilators.
For ‘closed book’ exams – phones/tablets and other mobile devices are not permitted in the assessment centre.
Candidates are expected maintain exam integrity and ethics by never releasing confidential examination materials or participate in fraudulent test-taking practices.
Examples include: 3.1 Giving coaching to any person who is to sit a Cyber Scheme exam within 2 years of having sat the same exam yourself. 3.2 Receiving coaching from a person who has sat a Cyber Scheme exam within 2 years of having sat the same exam.

Sanctions

A breach of this Code may lead to sanctions being applied, which may include, but are not limited to:

Assessments being immediately stopped and voided.
Any assessment awards could be either being withheld or revoked.
Revocation of any certification held by an individual with notification being given to the appropriate external bodies.

The Cyber Scheme reserve the right to wipe the hard drives of candidates at the end of an assessment, should a breach be proven. Any loss of personal or business data as a result of that is the responsibility of the candidate